Crouching + Vitals + Rad Damage Type
Player can now crouch using LCTRL. Crouch is currently on a toggle, might be worth adding an option in the gameplay section of the option menu to switch it to held down crouching. Also added Vital structure containing HP,Rads,Stamina and relevant resists. Finally added the file for Radiation Damage Type but haven't decide on what to do with it yet